Fix invalid named exports in property request page
